Revert plist real to float
Double value should be the correct type given the specification for the
binary plist format, however, we have a rounding issue when converting
to String in XML and ASCII plist format. The recommendation should be to
use BigDecimal but for now, we will just use float.

+++ subprojects/xcode-ide-kit/src/main/java/dev/nokee/xcode/PropertyListReader.java
@@ -48,5 +48,5 @@ public interface PropertyListReader extends Closeable {
 	boolean readBoolean();
 	String readString();
 	long readInteger();
-	double readReal();
+	float readReal();
 }
